requirements.txt
asgiref==3.6.0
attrs==22.1.0
bcrypt==4.0.1
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
jmespath==0.10.0
paramiko==2.12.0
pathspec==0.9.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
semantic-version==2.8.5
sqlparse==0.4.3
texttable==1.6.7
wcwidth==0.1.9
websocket-client==0.59.0
